moving API calls to services.
